adapt tests

+++ Tests/UnitTests/Core/Material/AverageMaterialsTests.cpp
@@ -24,10 +24,10 @@ using Acts::detail::combineSlabs;
 constexpr auto eps = std::numeric_limits<float>::epsilon();
 
 // vacuum w/ different thickness
-const Acts::MaterialSlab zeroVacuum = Acts::MaterialSlab(0.0f);
-const Acts::MaterialSlab unitVacuum = Acts::MaterialSlab(1.0f);
+const Acts::MaterialSlab zeroVacuum = Acts::MaterialSlab(0);
+const Acts::MaterialSlab unitVacuum = Acts::MaterialSlab(1);
 // same material corresponding to 0%, 1% and 100% radiation/interaction length
-const Acts::MaterialSlab zero(Acts::Test::makeSilicon(), 0.0f);
+const Acts::MaterialSlab zero(Acts::Test::makeSilicon(), 0);
 const Acts::MaterialSlab percent = Acts::Test::makePercentSlab();
 const Acts::MaterialSlab unit = Acts::Test::makeUnitSlab();
 
@@ -43,18 +43,18 @@ BOOST_AUTO_TEST_CASE(CombineSlabsVacuum) {
     auto slab = combineSlabs(zeroVacuum, zeroVacuum);
     BOOST_CHECK(!slab.isValid());
     BOOST_CHECK(!slab.material().isValid());
-    BOOST_CHECK_EQUAL(slab.thickness(), 0.0f);
-    BOOST_CHECK_EQUAL(slab.thicknessInX0(), 0.0f);
-    BOOST_CHECK_EQUAL(slab.thicknessInL0(), 0.0f);
+    BOOST_CHECK_EQUAL(slab.thickness(), 0.0);
+    BOOST_CHECK_EQUAL(slab.thicknessInX0(), 0.0);
+    BOOST_CHECK_EQUAL(slab.thicknessInL0(), 0.0);
   }
   // vacuum with unit thickness
   {
     auto slab = combineSlabs(unitVacuum, unitVacuum);
     BOOST_CHECK(!slab.isValid());
     BOOST_CHECK(!slab.material().isValid());
-    BOOST_CHECK_EQUAL(slab.thickness(), 2.0f);
-    BOOST_CHECK_EQUAL(slab.thicknessInX0(), 0.0f);
-    BOOST_CHECK_EQUAL(slab.thicknessInL0(), 0.0f);
+    BOOST_CHECK_EQUAL(slab.thickness(), 2.0);
+    BOOST_CHECK_EQUAL(slab.thicknessInX0(), 0.0);
+    BOOST_CHECK_EQUAL(slab.thicknessInL0(), 0.0);
   }
 }
 
@@ -186,13 +186,13 @@ BOOST_AUTO_TEST_CASE(CombineSlabsUnitZero) {
 
 BOOST_AUTO_TEST_CASE(CombineSlabsEqualThicknessVacuum) {
   const auto mat = Acts::Test::makeSilicon();
-  const auto slabMat = Acts::MaterialSlab(mat, 1.0f);
-  const auto slabVac = Acts::MaterialSlab(Acts::Material(), 1.0f);
+  const auto slabMat = Acts::MaterialSlab(mat, 1);
+  const auto slabVac = Acts::MaterialSlab(Acts::Material(), 1);
   {
     auto slab = combineSlabs(slabMat, slabVac);
     BOOST_CHECK(slab.isValid());
     BOOST_CHECK(slab.material().isValid());
-    BOOST_CHECK_EQUAL(slab.thickness(), 2.0f);
+    BOOST_CHECK_EQUAL(slab.thickness(), 2.0);
     BOOST_CHECK_EQUAL(slab.thicknessInX0(), slabMat.thicknessInX0());
     BOOST_CHECK_EQUAL(slab.thicknessInL0(), slabMat.thicknessInL0());
     // atomic mass and nuclear charge are per atom, adding any amount vacuum
@@ -200,18 +200,17 @@ BOOST_AUTO_TEST_CASE(CombineSlabsEqualThicknessVacuum) {
     BOOST_CHECK_EQUAL(slab.material().Ar(), mat.Ar());
     BOOST_CHECK_EQUAL(slab.material().Z(), 0.5 * mat.Z());
     // we have the same type of interactions just spread over twice the length
-    CHECK_CLOSE_REL(slab.material().X0(), 2.0f * mat.X0(), eps);
-    CHECK_CLOSE_REL(slab.material().L0(), 2.0f * mat.L0(), eps);
+    CHECK_CLOSE_REL(slab.material().X0(), 2 * mat.X0(), eps);
+    CHECK_CLOSE_REL(slab.material().L0(), 2 * mat.L0(), eps);
     // we have the same atoms just spread over more volume
-    BOOST_CHECK_EQUAL(slab.material().molarDensity(),
-                      0.5f * mat.molarDensity());
+    BOOST_CHECK_EQUAL(slab.material().molarDensity(), 0.5 * mat.molarDensity());
   }
   // reverse input order
   {
     auto slab = combineSlabs(slabVac, slabMat);
     BOOST_CHECK(slab.isValid());
     BOOST_CHECK(slab.material().isValid());
-    BOOST_CHECK_EQUAL(slab.thickness(), 2.0f);
+    BOOST_CHECK_EQUAL(slab.thickness(), 2.0);
     BOOST_CHECK_EQUAL(slab.thicknessInX0(), slabMat.thicknessInX0());
     BOOST_CHECK_EQUAL(slab.thicknessInL0(), slabMat.thicknessInL0());
     // atomic mass and nuclear charge are per atom, adding any amount vacuum
@@ -219,11 +218,10 @@ BOOST_AUTO_TEST_CASE(CombineSlabsEqualThicknessVacuum) {
     BOOST_CHECK_EQUAL(slab.material().Ar(), mat.Ar());
     BOOST_CHECK_EQUAL(slab.material().Z(), 0.5 * mat.Z());
     // we have the same type of interactions just spread over twice the length
-    CHECK_CLOSE_REL(slab.material().X0(), 2.0f * mat.X0(), eps);
-    CHECK_CLOSE_REL(slab.material().L0(), 2.0f * mat.L0(), eps);
+    CHECK_CLOSE_REL(slab.material().X0(), 2 * mat.X0(), eps);
+    CHECK_CLOSE_REL(slab.material().L0(), 2 * mat.L0(), eps);
     // we have the same atoms just spread over more volume
-    BOOST_CHECK_EQUAL(slab.material().molarDensity(),
-                      0.5f * mat.molarDensity());
+    BOOST_CHECK_EQUAL(slab.material().molarDensity(), 0.5 * mat.molarDensity());
   }
 }
 
@@ -232,51 +230,51 @@ BOOST_AUTO_TEST_CASE(CombineSlabsEqualThicknessVacuum) {
 BOOST_AUTO_TEST_CASE(CombineSlabs) {
   const auto mat0 = Acts::Material::fromMolarDensity(1, 1, 8, 12, 2);
   const auto mat1 = Acts::Material::fromMolarDensity(2, 2, 2, 6, 5);
-  const auto slabMat0 = Acts::MaterialSlab(mat0, 0.5f);
-  const auto slabMat1 = Acts::MaterialSlab(mat1, 1.0f);
+  const auto slabMat0 = Acts::MaterialSlab(mat0, 0.5);
+  const auto slabMat1 = Acts::MaterialSlab(mat1, 1.0);
   // verify derived quantities for the input slabs. these tests are not really
   // needed, but to show the input values for the tests below.
   BOOST_CHECK(slabMat0.isValid());
   BOOST_CHECK(slabMat0.material().isValid());
-  BOOST_CHECK_EQUAL(slabMat0.thickness(), 0.5f);
-  BOOST_CHECK_EQUAL(slabMat0.thicknessInX0(), 0.5f);
-  BOOST_CHECK_EQUAL(slabMat0.thicknessInL0(), 0.5f);
+  BOOST_CHECK_EQUAL(slabMat0.thickness(), 0.5);
+  BOOST_CHECK_EQUAL(slabMat0.thicknessInX0(), 0.5);
+  BOOST_CHECK_EQUAL(slabMat0.thicknessInL0(), 0.5);
   BOOST_CHECK(slabMat1.isValid());
   BOOST_CHECK(slabMat1.material().isValid());
-  BOOST_CHECK_EQUAL(slabMat1.thickness(), 1.0f);
-  BOOST_CHECK_EQUAL(slabMat1.thicknessInX0(), 0.5f);
-  BOOST_CHECK_EQUAL(slabMat1.thicknessInL0(), 0.5f);
+  BOOST_CHECK_EQUAL(slabMat1.thickness(), 1.0);
+  BOOST_CHECK_EQUAL(slabMat1.thicknessInX0(), 0.5);
+  BOOST_CHECK_EQUAL(slabMat1.thicknessInL0(), 0.5);
   // check combined slabs
   {
     auto slab = combineSlabs(slabMat0, slabMat1);
     BOOST_CHECK(slab.isValid());
     BOOST_CHECK(slab.material().isValid());
-    BOOST_CHECK_EQUAL(slab.thickness(), 1.5f);
-    BOOST_CHECK_EQUAL(slab.thicknessInX0(), 1.0f);
-    BOOST_CHECK_EQUAL(slab.thicknessInL0(), 1.0f);
-    BOOST_CHECK_EQUAL(slab.material().X0(), 1.5f);
-    BOOST_CHECK_EQUAL(slab.material().L0(), 1.5f);
-    BOOST_CHECK_EQUAL(slab.material().Ar(), 3.0f);
-    BOOST_CHECK_EQUAL(slab.material().Z(),
-                      static_cast<float>(
-                          exp((0.5 / 1.5) * log(12.0) + (1.0 / 1.5) * log(6))));
-    BOOST_CHECK_EQUAL(slab.material().molarDensity(), 4.0f);
+    BOOST_CHECK_EQUAL(slab.thickness(), 1.5);
+    BOOST_CHECK_EQUAL(slab.thicknessInX0(), 1.0);
+    BOOST_CHECK_EQUAL(slab.thicknessInL0(), 1.0);
+    BOOST_CHECK_EQUAL(slab.material().X0(), 1.5);
+    BOOST_CHECK_EQUAL(slab.material().L0(), 1.5);
+    BOOST_CHECK_EQUAL(slab.material().Ar(), 3.0);
+    BOOST_CHECK_EQUAL(
+        slab.material().Z(),
+        std::exp((0.5 / 1.5) * std::log(12.0) + (1.0 / 1.5) * std::log(6.0)));
+    BOOST_CHECK_EQUAL(slab.material().molarDensity(), 4.0);
   }
   // reverse input order
   {
     auto slab = combineSlabs(slabMat0, slabMat1);
     BOOST_CHECK(slab.isValid());
     BOOST_CHECK(slab.material().isValid());
-    BOOST_CHECK_EQUAL(slab.thickness(), 1.5f);
-    BOOST_CHECK_EQUAL(slab.thicknessInX0(), 1.0f);
-    BOOST_CHECK_EQUAL(slab.thicknessInL0(), 1.0f);
-    BOOST_CHECK_EQUAL(slab.material().X0(), 1.5f);
-    BOOST_CHECK_EQUAL(slab.material().L0(), 1.5f);
-    BOOST_CHECK_EQUAL(slab.material().Ar(), 3.0f);
-    BOOST_CHECK_EQUAL(slab.material().Z(),
-                      static_cast<float>(
-                          exp((0.5 / 1.5) * log(12.0) + (1.0 / 1.5) * log(6))));
-    BOOST_CHECK_EQUAL(slab.material().molarDensity(), 4.0f);
+    BOOST_CHECK_EQUAL(slab.thickness(), 1.5);
+    BOOST_CHECK_EQUAL(slab.thicknessInX0(), 1.0);
+    BOOST_CHECK_EQUAL(slab.thicknessInL0(), 1.0);
+    BOOST_CHECK_EQUAL(slab.material().X0(), 1.5);
+    BOOST_CHECK_EQUAL(slab.material().L0(), 1.5);
+    BOOST_CHECK_EQUAL(slab.material().Ar(), 3.0);
+    BOOST_CHECK_EQUAL(
+        slab.material().Z(),
+        std::exp((0.5 / 1.5) * std::log(12.0) + (1.0 / 1.5) * std::log(6.0)));
+    BOOST_CHECK_EQUAL(slab.material().molarDensity(), 4.0);
   }
 }
